v0.2.0 - Modular ft added to map parse validation

+++ lib/libft/ft_substr.c
@@ -20,9 +20,9 @@ char	*ft_substr(char const *s, unsigned int start, size_t len)
 	i = 0;
 	if (s == NULL)
 		return (NULL);
-	if (start > ft_strlen(s))
+	if (start > (size_t)ft_strlen(s))
 		return (ft_strdup(""));
-	if (len > (ft_strlen(s + start)))
+	if (len > (size_t)(ft_strlen(s + start)))
 		len = ft_strlen(s + start);
 	ptr = malloc((len + 1) * sizeof(char));
 	if (ptr == NULL)
